+++ packages/composer-layout/src/index.ts
@@ -5,6 +5,8 @@ export { DEFAULT_OVERLAY_CONTENT_MAX_FRACTION } from './types/layout';
 
 export { useViewportCategory } from './hooks/useViewportCategory';
 export { useKeyboardOpen } from './hooks/useKeyboardOpen';
+export { useViewportKeyboardState } from './hooks/useViewportKeyboardState';
+export type { ViewportKeyboardState } from './hooks/useViewportKeyboardState';
 export { useIsMobile } from './hooks/useIsMobile';
 export { useKeyboardOptionsSync } from './hooks/useKeyboardOptionsSync';